requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
bleach==6.0.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.2.2
Flask==2.0.2
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gotrue==0.5.4
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jinja2==3.1.4
keyring==24.1.1
logger==1.4
MarkupSafe==2.1.5
more-itertools==9.1.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pkginfo==1.10.0
postgrest==0.10.4
postgrest-py==0.10.4
proto-plus==1.24.0
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.18
pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-gitlab==3.15.0
python-semantic-release==7.32.1
readme-renderer==37.3
realtime==0.0.5
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.3.6
StrEnum==0.4.15
supabase==0.7.1
supafunc==0.2.2
tomlkit==0.12.5
tqdm==4.66.5
twilio==9.3.0
twine==3.8.0
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
webencodings==0.5.1
websockets==10.4
Werkzeug==2.0.3
yarl==1.9.4
zipp==3.15.0
gunicorn
